Show notes for each notebook successfully without redirection issues

+++ frontend/components/toolbar/toolbar.jsx
@@ -1,20 +1,29 @@
 import React from 'react';
+import { Redirect } from 'react-router-dom';
 
 export default class Toolbar extends React.Component{
   constructor(props){
     super(props);
-    this.delete = this.deleteItem.bind(this);
+    this.deleteItem = this.deleteItem.bind(this);
   };
 
   deleteItem() {
-
+    if (this.props.itemType === 'note') {
+      this.props.deleteNote(this.props.noteId);
+      this.props.history.push('/main/new');
+    }
   }
 
   render() {
-    return (
-      <div className="toolbar">
-        <div className="delete-button" onClick={this.deleteItem}>Delete</div>
-      </div>
-    )
+
+    if (this.props.display) {
+      return (
+        <div className="toolbar">
+          <div className="delete-button" onClick={this.deleteItem}>Delete</div>
+        </div>
+      )
+    } else {
+      return null;
+    }
   }
 }
